feat: Param Field Searchbar
- Added "value" field search term, to filter the fields to those with only that exact value.
- Added "range" field search term, to filter the fields to those with only a value between the two specified values.

+++ src/StudioCore/Editors/ParamEditor/Tools/DataInstanceFinder.cs
@@ -878,16 +878,16 @@ public static class DataFinderUtil
             case PARAMDEF.DefType.f32:
                 if (isRangeSearch)
                 {
-                    float floatStartVal;
-                    float floatEndVal;
+                    double floatStartVal;
+                    double floatEndVal;
 
-                    startSuccess = float.TryParse(startValue, out floatStartVal);
-                    endSuccess = float.TryParse(endValue, out floatEndVal);
+                    startSuccess = double.TryParse(startValue, out floatStartVal);
+                    endSuccess = double.TryParse(endValue, out floatEndVal);
 
                     if (startSuccess && endSuccess)
                     {
-                        if (((float)field.Value >= floatStartVal) &&
-                            ((float)field.Value <= floatEndVal))
+                        if (((double)field.Value >= floatStartVal) &&
+                            ((double)field.Value <= floatEndVal))
                         {
                             return (true, $"{field.Value}");
                         }
